+++ lib/danger/danger_core/executor.rb
@@ -1,3 +1,6 @@
+require "danger/danger_core/plugin_printer"
+require "danger/danger_core/dangerfile_printer"
+
 module Danger
   class Executor
     def run(env: nil,
@@ -6,10 +9,10 @@ module Danger
             base: nil,
             head: nil,
             dangerfile_path: nil,
-            danger_id: nil)
+            danger_id: nil,
+            verbose: false)
 
-      cork ||= Cork::Board.new(silent: false,
-                              verbose: false)
+      cork ||= Cork::Board.new(silent: false, verbose: false)
 
       # Could we find a CI source at all?
       unless EnvironmentManager.local_ci_source(ENV)
@@ -22,36 +25,55 @@ module Danger
         return
       end
 
-      # OK, then we can have some
+      # OK, then we can set ourselves up
       env ||= EnvironmentManager.new(ENV)
-      dm ||= Dangerfile.new(env, cork)
+      dm ||= dangerfile_for_path(dangerfile_path, env, cork)
 
-      dm.init_plugins
-
-      dm.env.fill_environment_vars
+      env.fill_environment_vars
 
       begin
-        dm.env.ensure_danger_branches_are_setup
+        env.ensure_danger_branches_are_setup
 
         # Offer the chance for a user to specify a branch through the command line
         ci_base = base || EnvironmentManager.danger_base_branch
         ci_head = head || EnvironmentManager.danger_head_branch
-        dm.env.scm.diff_for_folder(".", from: ci_base, to: ci_head)
+        env.scm.diff_for_folder(".", from: ci_base, to: ci_head)
 
         # Parse the local Dangerfile
         dm.parse(Pathname.new(dangerfile_path))
 
         post_results(dm, danger_id)
-        dm.print_results
+        print_results(env, cork) if verbose
       ensure
         dm.env.clean_up
       end
     end
 
+    # Gives you either a Dangerfile for Ruby, or a JS version
+    def dangerfile_for_path(path, env, cork)
+      klass = path.end_with?("js") ? DangerfileJS : Dangerfile
+      klass.new(env, cork)
+    end
+
+    # Prints out all the useful metadata
+    def print_results(env, cork)
+      # Print out the table of plugin metadata
+      plugin_printer = PluginPrinter.new(env.plugin_host)
+      plugin_printer.print_plugin_metadata(env, cork)
+
+      # Print out the results from the Dangerfile
+      messaging = env.plugin_host.core_plugins.first { |plugin| plugin.is_kind? DangerfileMessagingPlugin }
+      printer = DangerfilePrinter.new(messaging, cork)
+      printer.print_results
+    end
+
+    # Send the details to the request source
     def post_results(dm, danger_id)
       gh = dm.env.request_source
-      violations = dm.violation_report
-      status = dm.status_report
+      messaging = dm.env.plugin_host.core_plugins.first { |plugin| plugin.is_kind? DangerfileMessagingPlugin }
+
+      violations = messaging.violation_report
+      status = messaging.status_report
 
       gh.update_pull_request!(warnings: violations[:warnings], errors: violations[:errors], messages: violations[:messages], markdowns: status[:markdowns], danger_id: danger_id)
     end